+++ lib/service/Auth.dart
@@ -299,7 +299,7 @@ class Auth with Service implements NotificationsListener {
   Future<void> _syncProfilePiiDataIfNeed() async{
     if(isShibbolethLoggedIn){
       UserPiiData piiData = userPiiData;
-      if(piiData.updateFromAuthInfo(authInfo)){
+      if((piiData != null) && piiData.updateFromAuthInfo(authInfo)){
         storeUserPiiData(piiData);
       }
     }