+++ CHANGELOG.md
@@ -6,6 +6,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [2.4.3] - 2020-08-28
+### Fixed
+Fixed crash in Android [#61](https://github.com/rokwire/safer-illinois-app/issues/61)
+Fixed crash in Android [#68](https://github.com/rokwire/safer-illinois-app/issues/68)
+Fixed crash in Android when bluetooth is not turned on [#70](https://github.com/rokwire/safer-illinois-app/issues/70)
+Fixed crash in Android - do not start exposure client service if it's not running [#74](https://github.com/rokwire/safer-illinois-app/issues/74)
+
+### Updated
+SettingsNewHomePanel and related child panels moved to settings2 section, class names updated to indicate the different panel group.
+SettingsDebugPanel moved to debug section, MessagingPanel renamed to SettingsDebugMessagingPanel.
+
+### Deleted
+SettingsPrivacyCenterPanel (unused)
+Covid19OnBoardingLoginNetIdPanel (unused)
+Covid19OnBoardingLoginPhonePanel (unused)
+
+
 ## [2.4.2] - 2020-08-26
 ### Updated
 Fixed Xcode name in README.md (#1)